+++ fsl_mrs/tests/test_fslpy.py
@@ -18,7 +18,8 @@ pytest.importorskip('fsl.wrappers')
 
 from fsl.data.image import Image
 from fsl.wrappers import fsl_mrs, fsl_mrsi, fsl_mrs_preproc, fsl_mrs_preproc_edit, \
-                         fsl_mrs_proc, svs_segment, mrsi_segment
+                         fsl_mrs_proc, svs_segment, mrsi_segment, fsl_dynmrs, \
+                         basis2spec, fmrs_stats
 from fsl_mrs import read_basis
 from fsl_mrs.utils.validate_results import compare_folders
 
@@ -69,6 +70,23 @@ mrsi_segment_data = {
     'metab':    testsPath / 'testdata/fsl_mrsi/FID_Metab.nii.gz',
     'anat':     testsPath / 'testdata/mrsi_segment/T1.anat'}
 
+basis2spec_data = {
+    'basis':    testsPath / 'testdata/fsl_mrs/steam_basis_no_mm',
+    'reference': testsPath / 'testdata/fsl_mrs/metab.nii.gz'}
+
+fmrs_stats_data = {
+    'sim_results': [
+        testsPath / 'testdata/fmrs_tools/sim_fmrs/sub0/stim',
+        testsPath / 'testdata/fmrs_tools/sim_fmrs/sub1/stim',
+        testsPath / 'testdata/fmrs_tools/sim_fmrs/sub0/ctrl',
+        testsPath / 'testdata/fmrs_tools/sim_fmrs/sub1/ctrl'],
+    'fl_contrasts': testsPath / 'testdata/fmrs_tools/fl_contrasts.json',
+    'design_gm': testsPath / 'testdata/fmrs_tools/design_groupmean.mat',
+    'con_gm': testsPath / 'testdata/fmrs_tools/design_groupmean.con',
+    'design_paired': testsPath / 'testdata/fmrs_tools/design.mat',
+    'con_paired': testsPath / 'testdata/fmrs_tools/design.con',
+    'ftest': testsPath / 'testdata/fmrs_tools/design.fts',
+    'ftests': testsPath / 'testdata/fmrs_tools/design2.fts'}
 
 def _run_fsl_mrs_proc_cli(input_dir, output_dir):
     # 1. Combine water reference data for combination across dynamics
@@ -231,6 +249,47 @@ def _run_fsl_mrs_proc_wrapper(input_dir, output_dir, use_objects=False):
     fsl_mrs_proc.phase(file, output_dir, filename=filename, ppm=(4.6, 4.7))
 
 
+def _create_fsl_dynmrs_data(tmp_path):
+    import numpy as np
+    from fsl_mrs.core import basis
+    import fsl_mrs.utils.synthetic as syn
+    from fsl_mrs.core.nifti_mrs import gen_nifti_mrs
+
+    FID_basis1 = syn.syntheticFID(chemicalshift=[1,], amplitude=[1], noisecovariance=[[0]], damping=[3])
+    FID_basis2 = syn.syntheticFID(chemicalshift=[3,], amplitude=[1], noisecovariance=[[0]], damping=[3])
+    bset = basis.Basis(
+        np.stack((FID_basis1[0][0], FID_basis2[0][0]), axis=1),
+        ['Met1', 'Met2'],
+        axes=FID_basis1[2])
+    bset.basis_fwhm = [3 * np.pi, 3 * np.pi]
+
+    FID1 = syn.syntheticFID(chemicalshift=[1, 3], amplitude=[1, 1], noisecovariance=[[0.01]])
+    FID2 = syn.syntheticFID(chemicalshift=[1, 3], amplitude=[2, 2], noisecovariance=[[0.01]])
+    fid1 = FID1[0][0].reshape((1, 1, 1, 2048))
+    fid2 = FID2[0][0].reshape((1, 1, 1, 2048))
+    data = np.stack((fid1, fid2), axis=-1)
+    data = np.conj(data)
+
+    nmrs = gen_nifti_mrs(
+        data,
+        FID1[2].dwelltime,
+        FID1[2].SpectrometerFrequency,
+        dim_tags=['DIM_DYN', None, None])
+
+    time_var = np.arange(2)
+
+    # Save
+    data_path = tmp_path / 'data.nii.gz'
+    basis_path = tmp_path / 'basis'
+    tv_path = tmp_path / 'time_var.csv'
+
+    nmrs.save(data_path)
+    bset.save(basis_path)
+    np.savetxt(tv_path, time_var, delimiter=',')
+
+    return data_path, basis_path, tv_path
+
+
 @requires_cli('fsl_mrs')
 def test_fsl_mrs(tmp_path):
     cli_out     = tmp_path / 'fit_cli'
@@ -377,6 +436,7 @@ def test_fsl_mrs_preproc(tmp_path):
     ])
     assert (cli_out / 'metab.nii.gz').exists()
 
+    # fslpy call
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         fsl_mrs_preproc(
             data=preproc_data['metab'],
@@ -390,6 +450,7 @@ def test_fsl_mrs_preproc(tmp_path):
         )
     assert (wrapper_out / 'metab.nii.gz').exists()
 
+    # fslpy call with objects
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         fsl_mrs_preproc(
             data=Image(preproc_data['metab']),
@@ -428,6 +489,7 @@ def test_fsl_mrs_preproc_edit(tmp_path):
     ])
     assert (cli_out / 'diff.nii.gz').exists()
 
+    # fslpy call
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         fsl_mrs_preproc_edit(
             data=preproc_edit_data['metab'],
@@ -443,6 +505,7 @@ def test_fsl_mrs_preproc_edit(tmp_path):
         )
     assert (wrapper_out / 'diff.nii.gz').exists()
 
+    # fslpy call with objects
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         fsl_mrs_preproc_edit(
             data=Image(preproc_edit_data['metab']),
@@ -477,6 +540,7 @@ def test_svs_segment(tmp_path):
     ])
     assert (cli_out / 'segmentation.json').exists()
 
+    # fslpy call
     wrapper_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         svs_segment(
@@ -486,6 +550,7 @@ def test_svs_segment(tmp_path):
         )
     assert (wrapper_out / 'segmentation.json').exists()
 
+    # fslpy call with objects
     object_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         svs_segment(
@@ -514,6 +579,7 @@ def test_mrsi_segment(tmp_path):
     ])
     assert (cli_out / 'mrsi_seg_wm.nii.gz').exists()
 
+    # fslpy call
     wrapper_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         mrsi_segment(
@@ -523,6 +589,7 @@ def test_mrsi_segment(tmp_path):
         )
     assert (wrapper_out / 'mrsi_seg_wm.nii.gz').exists()
 
+    # fslpy call with objects
     object_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         mrsi_segment(
@@ -546,11 +613,13 @@ def test_fsl_mrs_proc(tmp_path):
     _run_fsl_mrs_proc_cli(preproc_data, cli_out)
     assert (cli_out / 'metab.nii.gz').exists()
 
+    # fslpy call
     wrapper_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         _run_fsl_mrs_proc_wrapper(preproc_data, wrapper_out)
     assert (wrapper_out / 'metab.nii.gz').exists()
 
+    # fslpy call with objects
     object_out.mkdir()
     with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
         _run_fsl_mrs_proc_wrapper(preproc_data, object_out, use_objects=True)
@@ -558,3 +627,162 @@ def test_fsl_mrs_proc(tmp_path):
 
     assert compare_folders(cli_out, wrapper_out, subdir=False)
     assert compare_folders(cli_out, object_out,  subdir=False)
+
+
+@requires_cli('fsl_dynmrs')
+def test_fsl_dynmrs(tmp_path):
+    cli_out     = tmp_path / 'dyn_cli'
+    wrapper_out = tmp_path / 'dyn_wrapper'
+    object_out  = tmp_path / 'dyn_object'
+
+    data_str, basis_str, tv_str = _create_fsl_dynmrs_data(tmp_path)
+    model_str = testsPath / 'testdata/dynamic/simple_linear_model.py'
+
+    subprocess.check_call([
+        'fsl_dynmrs',
+        '--data', data_str,
+        '--basis', basis_str,
+        '--dyn_config', model_str,
+        '--time_variables', tv_str,
+        '--baseline_order', '0',
+        '--output', cli_out,
+        '--report',
+    ])
+    assert cli_out.exists()
+    assert (cli_out / 'dyn_cov.csv').exists()
+    assert (cli_out / 'init_results.csv').exists()
+    assert (cli_out / 'dyn_results.csv').exists()
+    assert (cli_out / 'mapped_parameters.csv').exists()
+    assert (cli_out / 'free_parameters.csv').exists()
+    assert (cli_out / 'options.txt').exists()
+    assert (cli_out / 'report.html').exists()
+
+    # fslpy call
+    with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
+        fsl_dynmrs(
+            data=data_str,
+            basis=basis_str,
+            dyn_config=model_str,
+            time_variables=tv_str,
+            baseline_order='0',
+            output=wrapper_out,
+            report=True,
+        )
+    assert wrapper_out.exists()
+    assert (wrapper_out / 'dyn_cov.csv').exists()
+    assert (wrapper_out / 'init_results.csv').exists()
+    assert (wrapper_out / 'dyn_results.csv').exists()
+    assert (wrapper_out / 'mapped_parameters.csv').exists()
+    assert (wrapper_out / 'free_parameters.csv').exists()
+    assert (wrapper_out / 'options.txt').exists()
+    assert (wrapper_out / 'report.html').exists()
+
+    # fslpy call with objects
+    with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
+        fsl_dynmrs(
+            data=Image(data_str),
+            basis=read_basis(basis_str),
+            dyn_config=model_str,
+            time_variables=tv_str,
+            baseline_order='0',
+            output=object_out,
+            report=True,
+        )
+    assert object_out.exists()
+    assert (object_out / 'dyn_cov.csv').exists()
+    assert (object_out / 'init_results.csv').exists()
+    assert (object_out / 'dyn_results.csv').exists()
+    assert (object_out / 'mapped_parameters.csv').exists()
+    assert (object_out / 'free_parameters.csv').exists()
+    assert (object_out / 'options.txt').exists()
+    assert (object_out / 'report.html').exists()
+
+    assert compare_folders(cli_out, wrapper_out, subdir=True)
+    assert compare_folders(cli_out, object_out, subdir=True)
+
+
+@requires_cli('basis2spec')
+def test_basis2spec(tmp_path):
+    cli_out     = tmp_path / 'basis2spec_cli' / 'spectrum.nii.gz'
+    wrapper_out = tmp_path / 'basis2spec_wrapper' / 'spectrum.nii.gz'
+    object_out  = tmp_path / 'basis2spec_object' / 'spectrum.nii.gz'
+
+    subprocess.check_call([
+        'basis2spec',
+        '--basis', basis2spec_data['basis'],
+        '--reference', basis2spec_data['reference'],
+        '--output', cli_out,
+    ])
+
+    # fslpy call
+    with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
+        basis2spec(
+            basis=basis2spec_data['basis'],
+            reference=basis2spec_data['reference'],
+            output=wrapper_out,
+        )
+
+    # fslpy call with objects
+    with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
+        basis2spec(
+            basis=read_basis(basis2spec_data['basis']),
+            reference=Image(basis2spec_data['reference']),
+            output=object_out,
+        )
+
+    assert cli_out.exists()
+    assert wrapper_out.exists()
+    assert object_out.exists()
+    assert compare_folders(cli_out.parent, wrapper_out.parent, subdir=False)
+    assert compare_folders(cli_out.parent, object_out.parent,  subdir=False)
+
+
+@requires_cli('fmrs_stats')
+def test_fmrs_stats(tmp_path):
+    cli_out     = tmp_path / 'stats_cli'
+    wrapper_out = tmp_path / 'stats_wrapper'
+
+    results_list = tmp_path / 'results_list'
+    with open(results_list, 'w') as fp:
+        fp.writelines([str(x) + '\n' for x in fmrs_stats_data['sim_results']])
+
+    subprocess.check_call([
+        'fmrs_stats',
+        '--data', results_list,
+        '--output', cli_out,
+        '--fl-contrasts', fmrs_stats_data['fl_contrasts'],
+        '--combine', 'NAA', 'NAAG',
+        '--combine', 'Cr', 'PCr',
+        '--combine', 'PCh', 'GPC',
+        '--hl-design', fmrs_stats_data['design_gm'],
+        '--hl-contrasts', fmrs_stats_data['con_gm'],
+        '--hl-contrast-names', 'positive', 'negative',
+        '--overwrite',
+    ])
+    assert (cli_out / 'group_stats.csv').is_file()
+    assert (cli_out / '0_stim').is_dir()
+    assert (cli_out / '1_stim').is_dir()
+    assert (cli_out / '2_ctrl').is_dir()
+    assert (cli_out / '3_ctrl').is_dir()
+    assert (cli_out / '0_stim' / 'free_parameters.csv').is_file()
+
+    # fslpy call
+    with patch('fsl.utils.run.FSL_PREFIX', fsl_bin):
+        fmrs_stats(
+            data=str(results_list),
+            output=str(wrapper_out),
+            fl_contrasts=str(fmrs_stats_data['fl_contrasts']),
+            combine=[['NAA', 'NAAG'], ['Cr', 'PCr'], ['PCh', 'GPC']],
+            hl_design=str(fmrs_stats_data['design_gm']),
+            hl_contrasts=str(fmrs_stats_data['con_gm']),
+            hl_contrast_names=['positive', 'negative'],
+            overwrite=True,
+        )
+    assert (wrapper_out / 'group_stats.csv').is_file()
+    assert (wrapper_out / '0_stim').is_dir()
+    assert (wrapper_out / '1_stim').is_dir()
+    assert (wrapper_out / '2_ctrl').is_dir()
+    assert (wrapper_out / '3_ctrl').is_dir()
+    assert (wrapper_out / '0_stim' / 'free_parameters.csv').is_file()
+
+    assert compare_folders(cli_out, wrapper_out, subdir=True)